### 2022-06-10 #4466 @JeWaVe
Of course I will.
My citation comes from Google scholar, I'll try to find additional info.

Best regards

Regis.

Envoyé depuis ProtonMail mobile

-------- Message d'origine --------
Le 10 juin 2022 à 17:13, Kyle Niemeyer a écrit :

> Hi ***@***.***(https://github.com/JeWaVe), this pre-review issue is where we will find an editor and reviewers for your submission. Any suggestions you have for the latter would be welcome.
>
> Unfortunately, we do not have any editors in this area available to handle any additional submissions right now, so I will have to put your submission on our waitlist for the time being.
>
> In the meantime, could you check that missing DOI message pointed out above?
>
