### 2022-06-10 #4466 @JeWaVe
Paper is quite short as it looks to be the standard here. I can afford many explanations and benchmarks if needed though.

Best regards.

Envoyé depuis ProtonMail mobile

-------- Message d'origine --------
Le 10 juin 2022 à 18:42, The Open Journals editorial robot a écrit :

> Assigned! ***@***.***(https://github.com/diehlpk) is now the editor
>
