### 2022-06-18 #4466 @JeWaVe
@shumow @CaseyCarter Could you please help review this paper ?
It's a polynomial algrebra (over a discrete integral domain) library, full template.
@shumow It might interest you for the maths
@CaseyCarter It might be a good usage sample for the msvc c++ standard lib.
