### 2022-07-27 #4617 @editorialbot
[REVIEW]: Aerobus: a C++ template library for polynomials algebra over discrete integral domains
**Submitting author:** @JeWaVe (<a href="http://orcid.org/0000-0002-6291-3345">regis portalez</a>)
**Repository:** https://github.com/JeWaVe/aerobus
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.0
