### 2022-08-23 #4617 @JPenuchot
> Hi @JPenuchot, @balos1 how is your review going?

Hi Patrick, I have a few comments already:

- there is no documentation,
- the code is not commented,
- there are no instructions for installation (although it is a single header, having it specified in a documentation would be a good start),
- performance claims aren't backed by any benchmark/measure (nb: there is an assembly output example which might be a strong indicator, but isn't a performance measurement per se),
- the first commit was on the 3rd of june, and the last commit of the JOSS branch was on the 13th, which amounts to a 10 day effort and therefore doesn't nearly satisfy the substantial scholarly effort requirement of 3 months,
- [SLEEF](https://sleef.org/), which seems to implement the same functionalities, isn't cited anywhere and makes the statement of need fall short of arguments as of now. However I *do* understand the interest of having an alternative based on a generic implementation, which *could* be enough to justify the need for such a library (future-proofing for next architectures, or support for architectures not supported by SLEEF like RISC-V, etc...). Therefore the statement of need could be revised to something acceptable.

So in general the effort seems relatively small, and the project definitely falls short of any proper documentation and comprehensive comments, as well as performance measurements -preferably against the libraries cited by the author- to back the performance claims.

Regards,
Jules
