### 2022-08-23 #4617 @diehlpk
> * the first commit was on the 3rd of june, and the last commit of the JOSS branch was on the 13th, which amounts to a 10 day effort and therefore doesn't nearly satisfy the substantial scholarly effort requirement of 3 months,

@JeWaVe  could you please elaborate on that?
