### 2022-08-26 #4617 @balos1
Hi @diehlpk ,

So far, I have the same comments as @JPenuchot with my biggest concern being the lack of documentation or code commenting. The main branch does have installation instructions in the README (the joss brach does not), but that it is it. My second biggest concern is that this is work does not represent a substantial scholarly effort. I think further justification will be needed other than what has been provided so far.

Thanks,
Cody
