### 2022-09-02 #4617 @JeWaVe
Sure I will. I'm too busy with other projects right now but I will.

Envoyé depuis Proton Mail mobile

-------- Message d'origine --------
Le 2 sept. 2022 à 17:13, Patrick Diehl a écrit :

> Hi ***@***.***(https://github.com/JeWaVe),
>
> could you please provide more details on the time you spent on developing the code?
>
> What is the benefit of this code to the scientific community?
>
